Fix build for mac.
