+++ apps/desktop/src-tauri/src/utils/mod.rs
@@ -2,5 +2,6 @@ pub mod fs;
 pub mod icns;
 pub mod manifest;
 pub mod path;
+pub mod script;
 pub mod settings;
 pub mod setup;
