+++ apps/desktop/src/layouts/Default.astro
@@ -7,8 +7,6 @@ import { Toaster as SonnerToaster } from "@/components/ui/sonner";
 import PostHog from "@/components/posthog.astro";
 import GoogleAnalytics from "@/components/google-analytics.astro";
 import { ViewTransitions } from 'astro:transitions';
-import "element-plus/dist/index.css";
-import "element-plus/theme-chalk/dark/css-vars.css";
 
 interface Props {
   title: string;
